refactor: rename common database class
